requirements.txt
click==8.1.3
lxml==4.9.2
py-telegram-notifier==0.1.0.post2
pyquery==2.0.0
PyYAML==6.0
